src/camera_capture.py
"""
Camera Capture Module
Agent1担当: カメラからのリアルタイム映像取得

TODO for Agent1:
- CameraCaptureクラスの実装
- カメラ初期化処理
- フレーム取得メソッド
- エラーハンドリング
- tests/test_camera_capture.py の作成

詳細はAGENTS.mdを参照してください
"""

import cv2
import yaml
from typing import Optional, Tuple
import numpy as np


class CameraCapture:
    """
    固定カメラからのビデオキャプチャを管理するクラス

    Attributes:
        config (dict): カメラ設定
        cap (cv2.VideoCapture): OpenCVのVideoCaptureオブジェクト
    """

    def __init__(self, config: dict):
        """
        カメラキャプチャの初期化

        Args:
            config (dict): カメラ設定を含む辞書
                - device_id: カメラデバイスID
                - width: フレーム幅
                - height: フレーム高さ
                - fps: フレームレート
        """
        # TODO: 実装をここに追加
        pass

    def start(self) -> bool:
        """
        カメラキャプチャを開始

        Returns:
            bool: 成功した場合True
        """
        # TODO: 実装をここに追加
        pass

    def get_frame(self) -> Tuple[bool, Optional[np.ndarray]]:
        """
        カメラから1フレームを取得

        Returns:
            Tuple[bool, Optional[np.ndarray]]: (成功フラグ, フレーム画像)
        """
        # TODO: 実装をここに追加
        pass

    def stop(self):
        """
        カメラを停止してリソースを解放
        """
        # TODO: 実装をここに追加
        pass

    def is_opened(self) -> bool:
        """
        カメラが開いているか確認

        Returns:
            bool: カメラが開いている場合True
        """
        # TODO: 実装をここに追加
        pass


# テスト用のメイン関数
if __name__ == "__main__":
    # 設定ファイルを読み込んでテスト
    with open("../config.yaml", "r") as f:
        config = yaml.safe_load(f)

    camera = CameraCapture(config["camera"])
    # TODO: テストコードを追加
