requirements.txt
# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Data Processing
numpy==1.24.3

# Web Communication
websockets==12.0
requests==2.31.0

# Configuration
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
